+++ oidc-idp/src/main/java/cz/muni/ics/oidc/server/claims/ClaimSourceProduceContext.java
@@ -2,6 +2,7 @@ package cz.muni.ics.oidc.server.claims;
 
 import cz.muni.ics.oidc.models.RichUser;
 import cz.muni.ics.oidc.server.connectors.PerunConnector;
+import org.mitre.oauth2.model.ClientDetailsEntity;
 
 public class ClaimSourceProduceContext {
 
@@ -9,12 +10,14 @@ public class ClaimSourceProduceContext {
 	private final String sub;
 	private final RichUser richUser;
 	private final PerunConnector perunConnector;
+	private final ClientDetailsEntity client;
 
-	public ClaimSourceProduceContext(long perunUserId, String sub, RichUser richUser, PerunConnector perunConnector) {
+	public ClaimSourceProduceContext(long perunUserId, String sub, RichUser richUser, PerunConnector perunConnector, ClientDetailsEntity client) {
 		this.perunUserId = perunUserId;
 		this.sub = sub;
 		this.richUser = richUser;
 		this.perunConnector = perunConnector;
+		this.client = client;
 	}
 
 	public RichUser getRichUser() {
@@ -33,12 +36,15 @@ public class ClaimSourceProduceContext {
 		return perunConnector;
 	}
 
+	public ClientDetailsEntity getClient() {
+		return client;
+	}
+
 	@Override
 	public String toString() {
 		return "ClaimSourceProduceContext{" +
 				"perunUserId=" + perunUserId +
 				", sub='" + sub + '\'' +
-				//", richUser=" + richUser +
 				'}';
 	}
 }
