+++ oidc-idp/src/main/resources/logback.xml
@@ -50,10 +50,9 @@
 	<logger name="cz.muni.ics.oidc" level="${log.level}"/>
 	<logger name="cz.muni.ics.oidc.server.filters" level="error"/>
 	<logger name="cz.muni.ics.oidc.server.PerunUserInfo" level="info"/>
-	<logger name="cz.muni.ics.oidc.server.PerunUserInfoRepository" level="info"/>
+<!--	<logger name="cz.muni.ics.oidc.server.PerunUserInfoService" level="trace"/>-->
 	<logger name="cz.muni.ics.oidc.server.PerunTokenEnhancer" level="warn"/>
 	<logger name="cz.muni.ics.oidc.server.PerunAuthenticationUserDetailsService" level="info"/>
-	<logger name="cz.muni.ics.oidc.server.elixir.GA4GHClaimSource" level="trace"/>
-	<logger name="cz.muni.ics.oidc.server.connectors.impl.PerunConnectorRpc" level="debug"/>
+<!--	<logger name="cz.muni.ics.oidc.server.elixir.GA4GHClaimSource" level="trace"/>-->
 
 </configuration>
